pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ai_golf_coaches"
authors = ["Anders Brakestad"]
description = "AI golf coaches from youtube transcripts."
readme = "README.md"
version = "0.0.1"

[[tool.poetry.packages]]
include = "ai_golf_coaches"

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
pre-commit = "^4.3.0"
ruff = "^0.13.2"
mkdocs = "^1.6.1"
mkdocs-material = "^9.6.20"
mkdocs-awesome-nav = "^3.2.0"
google-api-python-client = "^2.183.0"
pydantic = "^2.11.9"
pydantic-settings = "^2.11.0"
youtube-transcript-api = "^1.2.2"
sentence-transformers = "^5.1.1"
nltk = "^3.9.1"
ollama = "^0.6.0"
python-dotenv = "^1.0.0"
requests = "^2.32.3"
tenacity = "^9.0.0"
PyYAML = "^6.0.2"

[tool.poetry.scripts]
aig = "ai_golf_coaches.cli:main"
